+++ src/__tests__/Calculator.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import {render, screen, fireEvent} from '@testing-library/react';
+import { render, screen, fireEvent } from '@testing-library/react';
 import Calculator from '../components/Calculator';
 
 test('calculator displays the whole calculator app', () => {
-  
-})
+
+});
