--- src/__tests__/InputText.test.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import {render, screen, fireEvent} from '@testing-library/react';
-import InputText from '../components/InputText';
-describe('InputTest Component', () => {
-  test('renders button with value', () => {
-   
-  });
-});
